requirements.txt
crowdin-api-client==1.13.0
requests==2.31.0
